docs: add the new paper
note that this was rejected from peer review in a non-disclosed journal due to not enough citations and not being "research paper"-like enough, which is fair in my opinion.

This should mostly be read as a general guide on how the tool works rather than being an official research paper.